fixed compile warning on 64 bits devices: Values of type 'NSUInteger' should not be used as format arguments; add an explicit cast to 'unsigned long' instead

+++ SPGooglePlacesAutocomplete/SPGooglePlacesAutocompleteQuery.m
@@ -39,7 +39,7 @@
                             [self.input stringByAddingPercentEscapesUsingEncoding:NSUTF8StringEncoding],
                             SPBooleanStringForBool(self.sensor), self.key];
     if (self.offset != NSNotFound) {
-        [url appendFormat:@"&offset=%u", self.offset];
+        [url appendFormat:@"&offset=%lu", (unsigned long)self.offset];
     }
     if (self.location.latitude != -1) {
         [url appendFormat:@"&location=%f,%f", self.location.latitude, self.location.longitude];
